Add a stand-alone replayer
The replayer can be used to reproduce crashes for fuzz targets using the
FuzzedDataProvider only given the raw crashing input. This will help
maintainers of OSS-Fuzz projects to reproduce issues even though they
have no access to the Java reproducers, which ClusterFuzz does not
store.

Compared to the full Jazzer driver, it has the advantage that it is a
Java library (with a native library dependency) and can thus be
published to Maven. Adding Windows support to the replayer will also be
much simpler than porting Jazzer itself to Windows.